git-svn-id: https://subversion.assembla.com/svn/kbasm/trunk/cpgf@14 5ea4cdcf-77fc-412b-903d-47c6f9a32d48

+++ include/cpgf/gmetaclass.h
@@ -555,7 +555,7 @@ public:
 			;
 	}
 
-	const GMetaConstructor * getConstructorByParamCount(unsigned int paramCount) const;
+	const GMetaConstructor * getConstructorByParamCount(size_t paramCount) const;
 	size_t getConstructorCount() const;
 	const GMetaConstructor * getConstructorAt(size_t index) const;
